ref: renamed fuzzing_stress to stress_test
