+++ src/tests/resistance/stress_test.rs
